Separate 'kr', 'en' in Roadmap
