+++ components/Roadmap/data/kr.ts
@@ -1,22 +1,6 @@
-export const RoadmapKey = {
-  PAST: 'PAST',
-  NOW: 'NOW',
-  FUTURE: 'FUTURE',
-  TEST: 'TEST',
-} as const;
+import { RoadmapKey, TotalRoadmap } from './types';
 
-export type Roadmap = {
-  title: string;
-  contents: string;
-  kind: 'DEV' | 'PLANNING' | 'MARKETING'; // 개발, 기획, 마케팅
-  dueDate?: string;
-};
-
-type TotalRoadmap = {
-  [key in keyof typeof RoadmapKey]: Roadmap[];
-};
-
-export const totalRoadmapKR: TotalRoadmap = {
+const roadmap: TotalRoadmap = {
   [RoadmapKey.PAST]: [
     {
       title: '빗썸 거래소 상장',
@@ -163,149 +147,4 @@ export const totalRoadmapKR: TotalRoadmap = {
   ],
 };
 
-export const totalRoadmapEN: TotalRoadmap = {
-  [RoadmapKey.PAST]: [
-    {
-      title: 'Bithumb Exchange Listing',
-      contents:
-        'To expand the liquidity of EL tokens, we listed EL tokens on the Bithumb Exchange, one of the largest exchanges in Korea.',
-      dueDate: '2020.03.25',
-      kind: 'MARKETING',
-    },
-    {
-      title: 'Gopax Exchange Listing',
-      contents:
-        'To expand the liquidity of EL tokens, we listed EL tokens on the Gopax Exchange, one of the largest exchanges in Korea.',
-      dueDate: '2020.11.20',
-      kind: 'MARKETING',
-    },
-    {
-      title: 'ELYFI Release',
-      contents:
-        'We launched DeFi, which provides financial services based on real-world asset tokens created through ELYSIA.',
-      dueDate: '2021.07.15',
-      kind: 'DEV',
-    },
-    {
-      title: 'On-chain Governance Launch',
-      contents:
-        'We introduced an on-chain governance system to strengthen the decentralization of ELYFI.',
-      dueDate: '2021.11.04',
-      kind: 'DEV',
-    },
-    {
-      title: 'MEXC Exchange Listing',
-      contents:
-        'To expand the liquidity of EL tokens, we listed EL tokens on the MEXC Exchange, one of the largest global exchanges.',
-      dueDate: '2021.12.01',
-      kind: 'MARKETING',
-    },
-    {
-      title: 'FBG Capital Investment Attraction',
-      contents:
-        'ELYSIA has raised $1 million in investment from FBG Capital, the largest venture capital in Asia.',
-      dueDate: '2022.01.13',
-      kind: 'MARKETING',
-    },
-    {
-      title: 'ELYFI BSC Mainnet Introduction',
-      contents:
-        'For the convenience of ELYFI users, we introduced the BSC mainnet, which is faster than Ethereum and has a lower gas cost.',
-      dueDate: '2022.01.20',
-      kind: 'DEV',
-    },
-    {
-      title: 'DAO LLC in Wyoming, USA, Establishment',
-      contents:
-        'To decentralize legal ownership of real-world assets and automate cash flows, we established the first DAO LLC in Wyoming, USA, from Asia.',
-      dueDate: '2022.02.07',
-      kind: 'PLANNING',
-    },
-    {
-      title: 'ELYFI US Real Estate Product Launch',
-      contents:
-        "To utilize the liquidity of the world's largest US real estate market, we launched a fractional investment product based on US real estate.",
-      dueDate: '2022.07.21',
-      kind: 'DEV',
-    },
-    {
-      title: ' Coinbase Custody Selection',
-      contents:
-        'ELYSIA has been selected for Coinbase custody, deposit and withdrawal of EL tokens are supported.',
-      dueDate: '2022.09.17',
-      kind: 'MARKETING',
-    },
-    {
-      title: 'EL Token Economy Improvement',
-      contents:
-        'We are improving ELYSIA’s token economy to strengthen decentralization and increase token value.',
-      dueDate: '2022.11.14',
-      kind: 'PLANNING',
-    },
-  ],
-  [RoadmapKey.NOW]: [
-    {
-      title: 'ELFI Token Economy Improvement',
-      contents:
-        'We are improving ELYFI’s token economy to strengthen decentralization and increase token value.',
-      dueDate: 'Expected by Q1 2023',
-      kind: 'PLANNING',
-    },
-    {
-      title: 'ELYSIA DAO LLC Articles of Incorporation Upgrade',
-      contents:
-        "To strengthen the decentralization of tokenization of real-world assets using the ELYSIA DAO LLC, we are upgrading the DAO LLC's articles of incorporation.",
-      dueDate: 'Expected by Q1 2023',
-      kind: 'PLANNING',
-    },
-    {
-      title: 'ELYFI v2 Release',
-      contents:
-        "Description: We are developing ELYFI v2 to complement ELYFI's contract.",
-      dueDate: 'Expected by Q1 2023',
-      kind: 'DEV',
-    },
-    {
-      title: 'Large Global Exchanges Listing',
-      contents:
-        'Description: To expand the liquidity of EL tokens, we are promoting listing on large global exchanges.',
-      dueDate: 'Expected by Q1 2023',
-      kind: 'MARKETING',
-    },
-  ],
-
-  [RoadmapKey.FUTURE]: [
-    {
-      title: 'Real-World Asset Tokenization Onboarding Page Release',
-      contents:
-        'We plan to launch an onboarding page where real-world asset owners can directly apply for tokenization.',
-      kind: 'DEV',
-    },
-    {
-      title: 'RWA Token Marketplace Introduction',
-      contents:
-        'We plan to launch a marketplace where RWA tokens based on real-world assets can be traded.',
-      kind: 'DEV',
-    },
-    {
-      title: 'ELYSIA Site Renewal',
-      contents:
-        'Description: We plan to renew the ELYSIA site for user convenience.',
-      kind: 'DEV',
-    },
-    {
-      title: 'ELYFI Site Renewal',
-      contents: 'We plan to renew the ELYFI site for user convenience.',
-      kind: 'DEV',
-    },
-  ],
-  [RoadmapKey.TEST]: [
-    {
-      title: 'Various Real Estate Asset Tokenization Logic Designs and Test',
-      contents:
-        'We are designing and testing logic for the commercialization of various real estates assets such as residential real estate, commercial real estate, and real estate PF.',
-      dueDate: 'Expected by Q4 2022',
-      kind: 'MARKETING',
-    },
-  ],
-};
+export default roadmap;
